chore: Electron boilerplate (#45)
* Electron Boilerplate

Hello world using electron, typescript, react, and webpack.
Start it using `npm start`

* Setup for packaging with electron-builder

Package with `npm run dist`. Packaged output will be written to /dist.

* Configure webpack to build tests

* Run tsc before webpack

On second thought, just run tsc first.

* Change window title

* Add React components for home screen and onboarding flow

* Add some basic frontend unit/integration tests

* Fix CI tests

* More unit tests

* Remove jest-setup.ts

Seems to work fine without it. I think this is taken care of already by including
jest-dom/extend-expect in jest.config.js

* Restore jest-setup.ts

It doesn't need to be included in the jest config,
but is still needed by typescript to compile tests.

* Use a more semantic jest-dom matcher

* Update package-lock.json

* Fix CI

* Implement requested changes from code review

* Move jestdom config lines to ui package

* Include tsx files in coverage reports

* Add spec for Home component

* Add typescript to devDeps

* Add callback from home for synchronize

* Add synchronize button

* Add Synchronize component

* Swap VError for Error

Having an issue with webpack and VError, or more specifically VError's
dependency on the node built-in Assert module. Going to just use the
generic JS Error class for the moment.

* Don't update a component if it is unmounted

* Add done and error state rendering to Synchronize

* Catch errors thrown by synchronizeWithCourier

Ironically, typescript doesn't let you declare the expected type of the error.

* Remove whitespace

* Fix linter errors

* Re-enable VError

* Prettier ignore /build

Reduce spurious failures of prettier on files we don't really look at anyway.

* Make async processes abortable w/ AbortController

Remove exceptions on mutability rules for tslint

Also adds a COMPLETE state to the Synchronize process to reduce
additional state tracking overhead in the frontend.

* Improve test coverage

* Use abortable-iterator npm module

Reduces branching in our code, thereby improving our branch test coverage score.

* Fix lint

* Remove random ;

* Set up for css loading

* Add some basic styles

* gitignore files generated by webpack

* Add HomeContent component

* Configure jest with mocks for css/file assets

* Setup for loading images

* Add some home styles

* Add some more shiny svgs

* Ignore webpacked svgs

* Add Inria Sans font

* Lint

* Refactor home component

Separate the rendering from the polling and state tracking code.

* Ignore webpack generated font files

* Add test for Status component

* Move jest-setup out of components

Because it's not a component, and so it won't appear amongst components
in test coverage reports.

* Refactor Synchronize component

Separate the rendering from the polling and state tracking code.

* Work on sync content plus button styles

* Fix home styling

* Add onboarding images

* Add test for SynchronizeStatus component

* Add onboarding styles

* Break up long lines

* Align buttons and dots in onboarding

* Fill in onboarding content

* Add content and layout styles for home

* Dial in synchronize content

* Style ordered list number markers in circles

* Simplify index component state

* Fix specs

* Add png logo

* Adjust heading size in Synchronize

* Add settings WIP

* Add menus to open app settings and other windows

* Allow reload with Cmd+R

* Add standard copy/paste keyboard shortcuts

Needed for normal text input interactions, e.g. copy-pasting a gateway

* Fix menu lint

* Style back buttons on settings screens

* Add confirmation checkbox

* Finish styling settings screens

* Validate gateway url and checkbox before submit

* Update gateway in settings component after succesfull migration

* Open the window a bit larger

* Mock electron to fix tests

Also move some ui-specific jest configs into the ui package

* Revert spurious change to global jest config

* Fix CI tests

* Rename react.tsx, refactor webpack config

react.tsx is now app.tsx, something slightly more meaningful
because i will soon add top-level tsx files for the about page and library lists.

Also extract the common parts of the webpack config for DRYness.

* Move app/package.json to src

This is essentially the source file for our app's package.json, which
is read by electron to determine the main entry point in a packaged
application. By keeping it in /src and letting webpack copy it over,
we can more easily clean and ignore the files generated by webpack in
/app.

* Generate the about page

* Rename electron.{ts|js} to main

By convention, because this is the entrypoint for the electron "main" process,
(in contrast to any electron "renderer" processes).

Also ensures more accurate matching when I search for this filename in my editor. :)

* Remove 'Hello World' flash before page loads

* Generate license report

* Add domain validity check in gateway editor

* Add test for libraries component

* Fix console warning

Warning: Each child in a list should have a unique "key" prop.

The key prop helps react keep the dom updated efficiently.

* Add test for AboutAwala component

* Add tests for gateway and gateway editor

* Add tests for Settings component

* Remove unused/empty state in Onboarding

* Make Onboarding default to the final step

We don't expect to get requests for content for invalid step numbers,
but if so, just default to the final step.

Co-authored-by: Gus Narea <gnarea@users.noreply.github.com>