Workaround for lerna and npm harmony
Use a relative path for the package.json version of daemon.
This allows npm to work correctly in the ui package.

Lerna ignores deps with directory names as versions, so also add a
call to npm link the daemon directory in the postinstall script.
This makes lerna bootstrap do the right thing for CI.