feat: Integrate UI and backend daemon (#75)
* Add dependency on the daemon package

* Webpack copies the daemon into the app

This  will copy the whole package folder, including both production and dev dependencies.
This is non-optimal. Probably we should get webpack to bundle the server, but an initial
attempt to do that resulted in some odd runtime errors, so taking the easy way out for now.

* Daemon integration WIP

Fork the daemon process
Get a token via IPC
Send the token to the UI
Pass the token to "ipc" calls (soon to be http API calls)

* Support closing or quitting from menu

* prettier / tslint

* Remove no-op error handler

* Fix dependency for lerna bootstrap

lerna bootstrap doesn't correctly handle local package dependencies if a relative
path is given instead of a version string.
See https://github.com/lerna/lerna/issues/2352

* Exclude daemon from license report

The report tool tries to look up the package in the registry, and fails to find a
matching version (0.0.0), because the registry's "daemon" is unrelated to our local
daemon package, so skip it.

* Update package-lock.json from newer npm version

Just updated my local npm for harmony with the version that lerna uses.

* Merge ui and daemon deps in Open Source Libraries

A quick and dirty js script to run the license report tool on both packages and merge the results.

* Workaround for lerna and npm harmony

Use a relative path for the package.json version of daemon.
This allows npm to work correctly in the ui package.

Lerna ignores deps with directory names as versions, so also add a
call to npm link the daemon directory in the postinstall script.
This makes lerna bootstrap do the right thing for CI.

* Integrate gateway address API using fetch

* Gateway address API fixes

* Add tests for gateway API calls

* Lint

* Test server API calls for method and headers

These requests should be GET/PUT with content type json.

* Update licenses

* Lint

* Update app display name

This field is used when building the packaged app as a user-facing nice display name.

* Update licenses

* Fix daemon launching

It wants to be run from the daemon directory to find the orm repo

* lint

* Make it possible to close and re-create the window

Clean up and refactor app setup and lifecycle management so we can close and open the window again
later, sending the current server auth token to any new window we create.

* lint

* Add option to open awala from application menu

* Add system tray menu

* Get connection status from ther server via websocket

* Update connectionStatus tests

Mocks the connect method from it-ws.

* Show connection status in system tray tooltip

* Courier sync websocket WIP

The server side isn't implemented yet, but we can already bootstrap this code based on the
connection status implementation. Just need to switch to using the courier sync endpoint.

* Ensure the UI exits if the socket closes or errors

* Update courier sync tests