Implement 'Try Again' button, Fixes #126
To reset a component we can change the value of the special `key` prop.
This signals to React that we want to throw away the old instance of
the component and create a fresh one, which will re-trigger any initial
setup, such as opening a new courier sync websocket connection in this case.